% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BdSize.R
\name{BdSize}
\alias{BdSize}
\title{Calculate the size of the SG building}
\usage{
BdSize(Bd_x, Bd_y)
}
\arguments{
\item{Bd_x:}{Vector containing min and max x coordinate of the building.}

\item{Bd_y:}{Vector containing min and max y coordinate of the building.}
}
\value{
A numeric value is returned, providing the size of the
        area of the building. Units are as input values are supplied.
}
\description{
Calculates the area of the SG building
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
