% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_foundation.R
\name{building_foundation}
\alias{building_foundation}
\title{Constructs a grid of the SG buidling foundation}
\usage{
building_foundation(Bdwall_ext_x, Bdwall_ext_y, Bdwall_ext_z, Bdbase_x,
  Bdbase_y, Bdbase_z, Pillar_x, Pillar_y, Pillar_z, grid_discretization)
}
\arguments{
\item{Bdwall_ext_x, y, z}{Numeric, giving the extent of the wall in x,y,z directions.}

\item{Bdbase_x, y, z}{Vector, giving min and max x,y,z values of the coordinates from the baseplate of the building.}

\item{Pillar_x, y, z}{Vector, giving min and max x,y,z values of the coordinates from the SG pillar within the building.}

\item{grid_discretization}{data.frame with columns (x,y,z), indicating the discretization in the corresponding direction.}
}
\value{
A data frame including x,y and z coordinates of walls, baseplate and SG pillar of the obervatory building.
}
\description{
Gridpoints for all points in space which belong to the
SG building. This includes walls, the baseplate and one SG pillar.
This foundation grid can later on be used to correct the gravity component grid,
in order to set mass changes equal zero at the foundation.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
