% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gravity_response.R
\name{calculate_gravity_response}
\alias{calculate_gravity_response}
\title{Calulate gravity response for 3d grids}
\usage{
calculate_gravity_response(gcomp_grid, mass_input)
}
\arguments{
\item{gcomp_grid}{Data.frame, containing the gravity component grid (x,y,z, gcomp).}

\item{mass_input}{Data.frame, containing columns structure $datetime, $value, $Depth.
Important: soil moisture values have to be in %, not integer values. Otherwise the order of the resulting gravity response is by a factor 100.}
}
\value{
Returns a data.frame containing columns of time information and the corresponding integrative gravity response.
}
\description{
Calculate integrative gravity response, resulting from the mass changes due to changes in soil moisture contents.
}
\details{
test
}
\examples{
missing 
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
