% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeshift.r
\name{ccf_events}
\alias{ccf_events}
\title{Event-based calulation of correlation coefficients, lagtimes and signal to noise ratios}
\usage{
ccf_events(event, event_name, mv_win, nmax, stl_flag = F, plots = T, ...)
}
\arguments{
\item{event}{vector of start and end date of event.}

\item{event_name}{character, name of event (mainly used for the plot).}

\item{mv_win}{length of window for moving average to smooth data in order to calculate signal to noise ratios.}

\item{nmax}{number of outputed results per correlation.}

\item{plots}{logical, should the results be plotted? (default is TRUE).}
}
\description{
So far hard-coded for SM data of wettzell SGnew TS
}
\details{
Output is a data.frame with correlation coeficients, lagtime and signal to soise ratios for each sensor.
}
\examples{

}
\references{
Marvin Reich (2015), mreich@gfz-potsdam.de
}
