% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeshift.r
\name{ccf_multievent}
\alias{ccf_multievent}
\title{Multi-event based ccf approach, including lagtimes and SNR}
\usage{
ccf_multievent(events, mv_win, nmax, limit_cor, correct_cor = T, limit_lag,
  correct_lag = T, plotting = T)
}
\arguments{
\item{events}{data.frame of events; each row limits one event, event[,1] = start dates, event[,2] = end dates, event[,3] = event name.}

\item{mv_win}{length of window for moving average to smooth data in order to calculate signal to noise ratios.}

\item{nmax}{number of outputs of correlation.}

\item{limit_cor}{threshold for correlation coefficient values; all values below limit_cor will be discarted.}

\item{correct_cor, correct_lag}{logical. should these quality filters be applied (TRUE, default) or only be included into statistics (done always).}

\item{limit_lag}{threshold for lagtime values; all values above limit_lag will be discarted.}

\item{plotting}{logical, should the results be plotted? (default is TRUE).}
}
\description{
Input events are filtered by SNR and optionally correlation and lagtime thresholds. The filtered datasets are then cross-correlated.

The resulting correlation coefficents, lagtimes and statistics are plotted and returned in form of a table (data frame).

So far hard-coded for SM data of wettzell SGnew TS
}
\details{
Output is a data.frame with mean values for correlation coeficients, lagtime and quality of signal to soise ratio.
}
\examples{

}
\references{
Marvin Reich (2015), mreich@gfz-potsdam.de
}
