% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeshift.r
\name{cclag}
\alias{cclag}
\title{Estimation of correlation coefficients and time lags of two timeseries (via cross-correlation )}
\usage{
cclag(data1, data2, nmax, normdata = T, mv_win = 96, ccfplot = F,
  resplot = F, sensor1 = "data1", sensor2 = "data2", ...)
}
\arguments{
\item{data1, data2}{Timeseries of type .zoo.}

\item{nmax}{Number of correlation values output.}

\item{mv_win}{window width used for signal to noise detection. value is numerical and represents data-timesteps.}

\item{ccfplot}{logical. Plots for each time series interval the cross-correlation plots. Default is F, own output plots are provided. Should be kept FALSE.}

\item{resplot}{logical. Plots an overall plot with results of correlation coefficients, lagtimes and signal to noise relationships of the complete datasets.}

\item{sensor1, sensor2}{names used for the input datasets for plotting and in the output table.}

\item{...}{Further parameters passed to internal functions.}

\item{norm}{logical. Use normalized data (default) or not.}
}
\description{
THIS IS A NEW VERSION..ojo!! has to be decided and merged afterwards!!!!

or replace old timeshift()-function !?

This function is based on standard time series decomposition and afterwards cross-correlation estimation between the two input datasets.

The main aim is to find dominant time shifts between two time series (e.g. soil moisture, climate data, precipitation stations, etc).
}
\details{
missing
}
\examples{
outside.cor = ccf.zoo(besidesBuilding$mux43_04,besidesBuilding$mux43_08,5,T,T,T,24,17521)

}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}
