% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funcs.R
\name{construct_vertices}
\alias{construct_vertices}
\title{Construct vertices for SG building foundation}
\usage{
construct_vertices(x_cords, y_cords, z_cords)
}
\arguments{
\item{x, y, z_cords}{Vector, holding min and max values of the x,y and z points which fix the polyheder.}
}
\description{
Following ptinpoly-packages, this function creates vertices
which can be used to correct gravity component grids for building structures
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
