% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funcs.R
\name{convert_demtodf}
\alias{convert_demtodf}
\title{Convert DEM to data.frame}
\usage{
convert_demtodf(dem, dem.info)
}
\arguments{
\item{grid_cords}{coordinates of the original hydrus mesh.}

\item{data_input}{data to be interpolated. in fomat of the hydrus mesh.}

\item{grid_discr}{vector of discretization of new grid in (x,y,z).}
}
\description{
interpolate DEM to necessary grid of corresponding layer
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
