% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gravity_response.R
\name{convert_gravity_response}
\alias{convert_gravity_response}
\title{Convert gravity data}
\usage{
convert_gravity_response(gravity_input, factor_hydScen_SGloc, factor_BdSize)
}
\arguments{
\item{gravity_input}{Data.frame, time series containing time information (datetime) and values of the gravity response from outside of the building.}

\item{factor_hydScen_SGloc}{Data.frame, time series having two columns: time information (datetime) and
the temporal dynamic factor for reduction due to hydrological scenario and SG position.}

\item{factor_BdSize}{Numeric, one value containing the reduction factor due to the influence of the size of the SG building.}
}
\value{
Returns a data.frame (time series), which has information about time (as the gravity_input time series) and
as value the converted gravity response for below the SG building. This converted gravity time series can now be 
used to reduce observed gravity data from this very SG building.
}
\description{
Converts gravity data from outside of the building to the gravity response
below the building. This is done multiplying with factors 1 & 2, both
obtained as a consequence of dominant hydrological scenario, SG position and SG building size.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
