% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_SGbuilding_foundation.R
\name{correct_SGbuilding_foundation}
\alias{correct_SGbuilding_foundation}
\title{Correct the gravity component grid for the SG building foundation}
\usage{
correct_SGbuilding_foundation(gravity_gcomp3d, Bdwall_ext_x, Bdwall_ext_y,
  Bdwall_ext_z, Bdbase_x, Bdbase_y, Bdbase_z, Pillar_x, Pillar_y, Pillar_z,
  grid_discretization)
}
\arguments{
\item{gravity_gcomp}{Data.frame, 4 columns: spatial information (x,y,z) and corresponding gravity component (value).}

\item{building_foundation}{Data.frame, 3 columns: spatial information (x,y,z) about the stucture of the foundation of the SG building.}
}
\value{
Returns a data.frame, consisting of 4 columns: 3 dimensions in space (x,y,Z) and value, holding the gravity component data.
}
\description{
In order to set all mass variations to zero where a firm and fix building can be found,
this function modifies the gravity component grid in the way that all gravity components laying physically
within the building foundations, are set to zero.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
