% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funcs.R
\name{decimalplaces}
\alias{decimalplaces}
\title{Calculate number of decimal places in a number}
\usage{
decimalplaces(x)
}
\arguments{
\item{x}{Numeric, the number to check decimal places.}
}
\description{

}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
