% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fill_gcompgrid}
\alias{fill_gcompgrid}
\title{test}
\usage{
fill_gcompgrid(g_grid, senloc, g_discr)
}
\arguments{
\item{g_grid}{grid for which to calculate gravity components. (generally derived from DEM of region in combination with hydrological modeling mesh/grid).}

\item{senloc}{coordinates of gravity sensor location. data.frame with x,y,z columns.}

\item{g_discr}{discretization of g_grid in differences in x,y,and z-direction. they have to be uniform steps!}

\item{test}{test}
}
\value{
test
}
\description{
test

Calculates the gravity components (100percent values) for a given DEM using a nested approach
}
\details{
missing

missing
}
\examples{
missing
example MISSING
}
\references{
Marvin Reich (2017), mreich@posteo.de

Marvin Reich (2014), mreich@gfz-potsdam.de
}
