% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.r
\name{filter_soildata}
\alias{filter_soildata}
\title{Filtering raw soil moisture data}
\usage{
filter_soildata(file.input, val.max = 1, val.min = 0, val.incr, val.decr,
  lag.max, per.max, per.min, lag.mean, k)
}
\arguments{
\item{file.input}{input dataset (zoo)}

\item{lag.max}{number of timesteps to consider for maximal increase / decrease}

\item{lag.mean}{period for moving mean calculation}

\item{k}{splits datasets at every k-th element of lag.mean}

\item{val.max/min}{cuts all data greater/smaller than this value}

\item{val.incr/decr}{maximum increase/decrease in absolute m³/m³ within timesteps lag.max}

\item{per.max/min}{value in percent for maximum / minimum deviation between actual value and a mean value of the time series}
}
\description{
Filtering raw soil moisture data with adjustable filter criterias.
}
\details{
The output will be as zoo-object.

Default values for val.max=1 and val.min=0.

Possible values for lag.mean are "hours", "days", "weeks", "months", "quarters" or "years".
}
\examples{
example MISSING
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}
