% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{forsberg_raw}
\alias{forsberg_raw}
\title{test}
\usage{
forsberg_raw(gama, w, xl, xr, yl, yr, Zint, Zend, xs, ys, zs, rho)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
