% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_comp_grid.R
\name{gravity_comp_grid}
\alias{gravity_comp_grid}
\title{Gravity component grid}
\usage{
gravity_comp_grid(surface, SG_coordinates, grid_discretization, grid_depth)
}
\arguments{
\item{surface}{this}

\item{SG_coordinates}{this}

\item{grid_discretization}{this}

\item{grid_depth}{this}
}
\value{
Returned is a data.frame with 4 columns.
3 for coordinates in space (x,y,z) and one with
the gravity component corresponding to this gridpoint.
}
\description{
Generates a grid of gravity components.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
