% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{grid2d_difs}
\alias{grid2d_difs}
\title{Calculate delta values of 2 grids at different time steps}
\usage{
grid2d_difs(data1, data2, data3 = NA, timestep, valmin, valmax)
}
\arguments{
\item{data1, data2, data3}{input data. has to be in the format XX.}

\item{timestep}{vector of timesteps to be considered for calculations.}

\item{valmax, valmin}{values where to cut off resulting differces, leaving a data-area of special consideration}
}
\description{
2D
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
