% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{grid3d_plot}
\alias{grid3d_plot}
\title{plot 3D grid}
\usage{
grid3d_plot(input, type)
}
\arguments{
\item{data1, data2, data3}{input data. has to be in the format XX.}

\item{timestep}{vector of timesteps to be considered for calculations.}

\item{valmax, valmin}{values where to cut off resulting differces, leaving a data-area of special consideration}
}
\description{
or differences between timesteps or absolute values each timestep
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
