% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{grid_stats_2d}
\alias{grid_stats_2d}
\title{compute grouped standard deviation, range, max, mean from values of 2D grid}
\usage{
grid_stats_2d(input, method)
}
\arguments{
\item{data1, data2, data3}{input data. has to be in the format XX.}

\item{timestep}{vector of timesteps to be considered for calculations.}

\item{valmax, valmin}{values where to cut off resulting differces, leaving a data-area of special consideration}
}
\description{
using sorting and a moving window average
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
