% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{hydrus_dem}
\alias{hydrus_dem}
\title{Data preparation for HYDRUS}
\usage{
hydrus_dem(inputpath, inputfile, dz, limit.area)
}
\arguments{
\item{inputpath}{DEM-data (realtive or absolute paths)}

\item{inputfile}{data.frame containing information about the DEM (row wise): columns, rows, starting x, starting y, lengths of one cellsize}

\item{dz}{vector containing the thickness of each subsurface layer}

\item{limit.area}{vector of coordinates declaring the area of interest within the inputfile (DEM). Structure: X(left), Y(top), X(right), Y(bottom).}
}
\description{
dem TS
}
\details{
Both DEM-data and its information file can be created using readDEM().
}
\examples{
missing 
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}
