% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{hydrus_lysidrain_ben}
\alias{hydrus_lysidrain_ben}
\title{Data preparation for HYDRUS}
\usage{
hydrus_lysidrain_ben(date.in, date.out, freq = "hour", area = 1,
  winlength = 10, aszoo = F)
}
\arguments{
\item{date.in}{starting date (POSIXct)}

\item{date.out}{ending date (POSIXct)}

\item{freq}{frequency of data. Possible values are "day", "hour", "min", "sec".}

\item{area}{bottom surface area of the lysimeter. In m².}

\item{aszoo}{output either in data.frame (aszoo=F) or zoo (aszoo=T). Default value is FALSE.}
}
\description{
Create lysimeter drainage time series
}
\details{
The outout will be given in meters drained water column per timestep (differences).
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}
