% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{hydrus_precip}
\alias{hydrus_precip}
\title{Data preparation for HYDRUS}
\usage{
hydrus_precip(date.in, date.out, freq = "hour", aszoo = F)
}
\arguments{
\item{date.in}{starting date (POSIXct)}

\item{date.out}{ending date (POSIXct)}

\item{freq}{frequency of data. Possible values are "day", "hour", "min", "sec".}

\item{aszoo}{output either in data.frame (aszoo=F) or zoo (aszoo=T). Default value is FALSE.}
}
\description{
Create precipitation time series
}
\details{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}
