% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{ModOut_animate}
\alias{ModOut_animate}
\title{make a sequence (movie) from model results}
\usage{
ModOut_animate(input1, input2, counter, type, outputname, valmax, valmin)
}
\arguments{
\item{valmax, valmin}{values where to cut off resulting differces, leaving a data-area of special consideration}

\item{data1, data2, data3}{input data. has to be in the format XX.}

\item{timestep}{vector of timesteps to be considered for calculations.}
}
\description{
or differences between timesteps or absolute values each timestep
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
