% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{nodestoFIXgrid_2d}
\alias{nodestoFIXgrid_2d}
\title{Interpolate 2D nodes to regular, already FIXED grid}
\usage{
nodestoFIXgrid_2d(fixgrid, data_input, nintpmax)
}
\arguments{
\item{data_input}{data to be interpolated. in fomat of the hydrus mesh.}

\item{grid_cords}{coordinates of the original hydrus mesh.}

\item{grid_discr}{vector of discretization of new grid in (x,y,z).}
}
\description{
interpolate hydrus mesh output to regular grid
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
