% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{nodestogrid_1d}
\alias{nodestogrid_1d}
\title{Interpolate 1D nodes to regular grid}
\usage{
nodestogrid_1d(data_input, grid_discr, depth_split)
}
\arguments{
\item{data_input}{data to be interpolated. in fomat of the hydrus mesh.}

\item{grid_discr}{vector of discretization of new grid in (x,y,z).}

\item{grid_cords}{coordinates of the original hydrus mesh.}
}
\description{
interpolate hydrus mesh output to regular grid
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
