% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{nodestogrid_2d_krige}
\alias{nodestogrid_2d_krige}
\title{Interpolate 2D nodes to regular grid, using kriging}
\usage{
nodestogrid_2d_krige(grid_cords, data_input, grid_discr, depth_split, tstart)
}
\arguments{
\item{grid_cords}{coordinates of the original hydrus mesh.}

\item{data_input}{data to be interpolated. in fomat of the hydrus mesh.}

\item{grid_discr}{vector of discretization of new grid in (x,y,z).}
}
\description{
interpolate hydrus mesh output to regular grid
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
