% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{nodestogridPP}
\alias{nodestogridPP}
\title{PARALLEL Interpolate 3D nodes to regular grid}
\usage{
nodestogridPP(grid_cords, data_input, grid_discr, depth_split)
}
\arguments{
\item{grid_cords}{coordinates of the original hydrus mesh.}

\item{data_input}{data to be interpolated. in fomat of the hydrus mesh.}

\item{grid_discr}{vector of discretization of new grid in (x,y,z).}
}
\description{
interpolate hydrus mesh output to regular grid
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
