% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{obsNode_hydrus}
\alias{obsNode_hydrus}
\title{Select time series of one modeled node}
\usage{
obsNode_hydrus(nodal_info_in, loc_hor, loc_vert, sensorname)
}
\arguments{
\item{nodal_info_in}{input datasets. should be a data.frame}

\item{loc_hor}{horizontal coordinate (in model coordintaes)}

\item{loc_vert}{vertical coordinate (in model coordinates)}

\item{sensorname}{name of sensor (for plotting and structuring)}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}
