% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{obsNodePlot}
\alias{obsNodePlot}
\title{Plot modeled nodes and observed soil moisture sensors}
\usage{
obsNodePlot(filename, normdata = T, dim2d = T, time_start, time_end,
  filtersen = F, hydrusnodes)
}
\arguments{
\item{filename}{filename of time series of the observation nodes (see details).}

\item{normdata}{logical. compare absolute (FALSE) or normalized values (TRUE; default).}

\item{dim2d}{TRUE for 2d, FALSE for 3D input}
}
\description{
plot time series of modeled observation nodes and measured soil moisture. This works with both hydrus 2D and 3D models.
}
\details{
time series of observation nodes can be extracted and saved using function read_obsNode2d() or read_obsNode3d()
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
