% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gcomp_grid.R
\name{plot_gcomp_grid}
\alias{plot_gcomp_grid}
\title{Plot transect of gravity component grid}
\usage{
plot_gcomp_grid(grid_input, yloc, output_dir, grid_discretization, ...)
}
\arguments{
\item{grid_input}{Data.frame, containing spatial coordinates in the columns (x, y, z).}

\item{yloc}{Numeric, coordinate of the y-axis, along which the transect will be constructed.}

\item{output_dir}{a}
}
\value{
Returns a ggplot figure. The plot is also saved to the specified output directory.
}
\description{
Plots a transect of a supplied input grid, meaning that a 3d grid is reduced to a 2d
transect on basis of a given axis (coordinate).
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
