% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{plot_hydrus1d_InOutput}
\alias{plot_hydrus1d_InOutput}
\title{Plot timeseries input / output : Hydrus 1D}
\usage{
plot_hydrus1d_InOutput(input_ts, tlevel_out, obsNode, timesteps, vertical_nodes,
  t_printout, tprint_show = 10)
}
\arguments{
\item{timesteps}{number of modeled timesteps}

\item{vertical_nodes}{mumber of nodes used in the vertical model discretisation}

\item{t_printout}{number of model printout times}

\item{folder}{foldername of project to read}

\item{plotting}{indicate if standard parameters should be plotted (TRUE); default is FALSE}

\item{datetime}{vector of timestamps of the modeled timeseries. if not provided output time will be in counts of timesteps}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}
