% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts_data.R
\name{plot_ts_data}
\alias{plot_ts_data}
\title{Plot multiply time series with one same basis}
\usage{
plot_ts_data(gravity_obs = NA, gravity_outside, gravity_below,
  gravity_reduced = NA, input_dir = dir_input, output_dir = dir_output,
  dat_tsf = 7, ...)
}
\arguments{
\item{gravity_obs}{Character string with the filename of the gravity observation data to be used.}

\item{gravity_outside}{Data.frame, containing gravity response corresponding to mass variations from outside of the buildling.
It has 2 columns (datatime and value).}

\item{gravity_below}{Data.frame, containing gravity response corresponding to mass variations from below of the buildling.
It has 2 columns (datatime and value).}

\item{gravity_reduced}{Data.frame, containing the reduced gravity data (gravity observation data corrected for mass variations from below the SG building).
It has 2 columns (datatime and value).}

\item{input_dir}{Character string, containing the path to input file "gravity_obs".}

\item{output_dir}{Character string, containing the path to the directory where to save the plot.}
}
\value{
Returns NULL as object. It creates a plot of all input time series.
This plot is shown on screen as well as saved to the specified output folder.
}
\description{
Plot all gravity time series used and calulated throught this reduction procedure in one joined plot.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
