% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_formulas.R
\name{pointmass}
\alias{pointmass}
\title{test}
\usage{
pointmass(gama, zp, zs, dx, dy, dz, rad, w, rho)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
