% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{predict_umbrellaspace}
\alias{predict_umbrellaspace}
\title{estimate function for umbrella space}
\usage{
predict_umbrellaspace(inputdata, displ_x, displ_y, polyorder, fitgeometry,
  plotting = T)
}
\arguments{
\item{data1, data2, data3}{input data. has to be in the format XX.}

\item{timestep}{vector of timesteps to be considered for calculations.}

\item{valmax, valmin}{values where to cut off resulting differces, leaving a data-area of special consideration}
}
\description{
using lm()
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
