% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{read_3dgrid}
\alias{read_3dgrid}
\title{Read Hydrus 3D grid coordinates}
\usage{
read_3dgrid()
}
\arguments{
\item{NA}{no parameter has to be inputed}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
