% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_csv}
\alias{read_csv}
\title{Read .csv files}
\usage{
read_csv(data_in, data_dir, spat_col = c(NA, NA, 2), dat_col = 3, ...)
}
\arguments{
\item{data_in}{Vector, containing the filename of the data to read.}

\item{...}{additional parameters for reading .csv lines. (sep, dec, etc.). 
See documentation of read.csv() for a detailed list of options.}

\item{input_dir}{Vector, containing the directory of the file.}
}
\value{
Returns a data.frame, consisting of a time series (2 columns, time info and data).
}
\description{
Reads .csv files and stores it as a data.frame.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
