% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{read_hydrus1d}
\alias{read_hydrus1d}
\title{Read Hydrus 1D output data}
\usage{
read_hydrus1d(folder, timesteps, vertical_nodes, t_printout, plotting = F,
  timestamps, tprint_show = 10)
}
\arguments{
\item{folder}{path to hydrus project folder}

\item{timesteps}{number of modeled timesteps}

\item{vertical_nodes}{mumber of nodes used in the vertical model discretisation}

\item{t_printout}{number of model printout times}

\item{plotting}{indicate if standard parameters should be plotted (TRUE); default is FALSE}

\item{timestamps}{vector of timestamps of the modeled timeseries. if not provided output time will be in counts of timesteps}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}
