% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{read_hydrus2d}
\alias{read_hydrus2d}
\title{Read Hydrus 2D output data}
\usage{
read_hydrus2d(folder, timesteps, factorforprintout, plotting = F, timestamps,
  profile_loc, vertical = T)
}
\arguments{
\item{folder}{foldername of project to read}

\item{timesteps}{number of modeled timesteps}

\item{plotting}{indicate if standard parameters should be plotted (TRUE); default is FALSE}

\item{timestamps}{vector of timestamps of the modeled timeseries. if not provided output time will be in counts of timesteps}

\item{t_printout}{string of dates of time series where informations are printed}

\item{px}{number of horizontal node, where the vertical profile should be analyzed}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}
