% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{read_hydrus3d}
\alias{read_hydrus3d}
\title{Read Hydrus 3D output data}
\usage{
read_hydrus3d(folder, timesteps, t_printout, plotting = F, timestamps, px)
}
\arguments{
\item{folder}{foldername of project to read}

\item{timesteps}{number of modeled timesteps}

\item{t_printout}{number of model printout times}

\item{plotting}{indicate if standard parameters should be plotted (TRUE); default is FALSE}

\item{timestamps}{vector of timestamps of the modeled timeseries. if not provided output time will be in counts of timesteps}

\item{px}{number of horizontal node, where the vertical profile should be analyzed}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}
