% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{read_hydrus3d_data}
\alias{read_hydrus3d_data}
\title{Read Hydrus 3D theta / head data}
\usage{
read_hydrus3d_data(folder, timestamps, datatype = "moisture")
}
\arguments{
\item{folder}{foldername of project to read}

\item{timestamps}{vector of timestamps of the modeled timeseries.}

\item{datatype}{character. possible values are: "moisture", "phead".}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
