% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{read_hydrus3d_dataBD}
\alias{read_hydrus3d_dataBD}
\title{Read Hydrus 3D theta / head data version FOR BIG DATA}
\usage{
read_hydrus3d_dataBD(filenames, timestamps, datatype = "moisture")
}
\arguments{
\item{filenames}{character vector of filenames for model output (ordered): time, soil moisture, pressure head.}

\item{timestamps}{vector of timestamps of the modeled timeseries.}

\item{datatype}{character. possible values are: "moisture", "phead".}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
