% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{read_obsNode3d}
\alias{read_obsNode3d}
\title{Read Hydrus 3D observation node data}
\usage{
read_obsNode3d(folder, realTime = T, startdate, plotting = F)
}
\arguments{
\item{folder}{foldername of project to read}

\item{realTime}{logical. Output in UTC (TRUE; default) or model time steps (FALSE).}

\item{startdate}{starting date for output in UTC. Format is POSIXct.}

\item{plotting}{do you want node time series plotted (default is FALSE).}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}
