% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_tsf}
\alias{read_tsf}
\title{Read .tsf files}
\usage{
read_tsf(data_in, data_dir, dat_col = 7, ...)
}
\arguments{
\item{data_in}{Vector, containing the filename of the data to read.
Please note that the .tsf file should end with regular data and NOT a special statement.}

\item{input_dir}{Vector, containing the directory of the file.}
}
\value{
Returns a data.frame, consisting of a time series (2 columns, time info and data).
}
\description{
Reads .tsf files and stores it as a data.frame.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
