% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_gravity.R
\name{reduce_gravity}
\alias{reduce_gravity}
\title{Reduce observed gravity data}
\usage{
reduce_gravity(gravity_obs, gravity_below, input_dir, ...)
}
\arguments{
\item{gravity_obs}{Vector, containing the filename of the observed gravity data, which is to be reduced.}

\item{gravity_below}{Data.frame, containing time (datatime) and gravity response data (value).}

\item{input_dir}{Vector, containing the directory of the gravity_data-file.}
}
\value{
Returns a data.frame, consisting of a time series (2 columns, time info and data)
of the reduced gravity response. This means that observed gravity data was corrected (subtracted) by
the response due to mass variation from below observatory buildings.
This way improving (or cleaning) the observed gravtiy signal from near-field local hydrology due to the umbrella effect.
}
\description{
Reduces (corrects) observed gravity data by mass variations occuring and originating below SG observatory buildings.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
