% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_BdSize.R
\name{reduction_BdSize}
\alias{reduction_BdSize}
\title{Receive reduction factor based on the dimension of the SG building}
\usage{
reduction_BdSize(SG_BdSize)
}
\arguments{
\item{MeanSoilMoisture}{Data.frame, time series of the mean soil moisture content in the complete space outside of the observatory building.
It needs 2 columns: one for time information (datetime) and one with the soil moisture data (value).}
}
\value{
Returns a numerical values, which represents the correction factor based on the dimension of the SG building.
This can later be used to adjust the modeled gravity response from outside of the building.
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
