% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_hydScen_SGloc.R
\name{reduction_hydScen_SGloc}
\alias{reduction_hydScen_SGloc}
\title{Receive reduction factor based on hydrological conditon and position of SG}
\usage{
reduction_hydScen_SGloc(Scenario, SGlocation, VertLimit, MeanSoilMoisture)
}
\arguments{
\item{Scenario}{Character vector, defining the dominant hydrological condition found at or around the observatory building.
For possible options please read the instructions.}

\item{SGlocation}{Character string, describing the position of the SG within its building.
For possible options please read the instructions.}

\item{VertLimit}{Numeric value of the vertical extent to be considered for the reduction. Values are in meters and have to be positive.}

\item{MeanSoilMoisture}{Data.frame, time series of the mean soil moisture content in the complete space outside of the observatory building.
It needs 2 columns: one for time information (datetime) and one with the soil moisture data (value).}
}
\value{
Returns a time series of numerical values, which represents the dynamic correction factors based on a dominant hydrological condition
and the in-house position of the SG.
This can later be used to adjust the modeled gravity response from outside of the building.
}
\description{
A dynamic correction ratio is created, based on mean soil moisture contents next to the observatory building and the chosen conditons
of a dominant hydrological scenario as well as the position of the SG within its building.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
