% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.r
\name{snr_filter}
\alias{snr_filter}
\title{Signal to noise filtering}
\usage{
snr_filter(data.in, mv_win, rmNA = T, plotting = F)
}
\arguments{
\item{data.in}{input time series dataset (zoo)}

\item{mv_win}{width of moving average window, in observation time steps (no explicit time unit)}

\item{rmNA}{logical. should NA values be remove within the forming of mean values? default is TRUE}

\item{plotting}{logical. if true (not default) datasets are plotted}
}
\description{
Filters input data based on signal to noise ratios (SNR).
}
\details{
The output is a time series: signal =  the input time series (raw data) - noise.
}
\examples{
example MISSING

}
\references{
Marvin Reich (2015), mreich@gfz-potsdam.de
}
