% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilMoisture_grid3d.R
\name{SoilMoisture_grid3d}
\alias{SoilMoisture_grid3d}
\title{Soil moisture grid (3d)}
\usage{
SoilMoisture_grid3d(grid_domain = gravity_component_grid3d,
  soilMoisture_input = soilMoisture_input_file, input_dir = dir_input, ...)
}
\arguments{
\item{grid_domain}{Data.frame, constructed with minimum 3 columns (x,y,z) which describe spatial dimensions of 3d grid.}

\item{soilMoisture_input}{Name of file, containing the soil moisture time series.
This file has to contain minimal 2 columns (datetime and z) but can optionally also contain (x and/or y), if a higher dimensional soil moisture dataset is available.}

\item{input_dir}{Directory where}
}
\value{
Returns a data.frame with 5 columns.
Information about time (datetime), spatial information (x,y,z) and value which is the soil moisture data.
}
\description{
Generates a grid in 3d, containing time series
at each point of soil moisture.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
