% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrus.r
\name{umbrella_points_cluster}
\alias{umbrella_points_cluster}
\title{estimate limiting points of umbrella space / curve for ALL y-profiles with SD of soil moisture point time series}
\usage{
umbrella_points_cluster(data1, clusternum, counters)
}
\arguments{
\item{data1, data2, data3}{input data. has to be in the format XX.}

\item{timestep}{vector of timesteps to be considered for calculations.}

\item{valmax, valmin}{values where to cut off resulting differces, leaving a data-area of special consideration}
}
\description{
scanning area from beneath, looking for depth with smallest difference
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2016), mreich@gfz-potsdam.de
}
