% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funcs.R
\name{zootodf}
\alias{zootodf}
\title{Convert zoo-object to data frame}
\usage{
zootodf(value)
}
\arguments{
\item{value}{must be a zoo-object}
}
\description{
Converting zoo-object to data frame with proper indexing of time column and column names
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
