% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_comp_grid.R
\name{select_gMethod_Edges}
\alias{select_gMethod_Edges}
\title{Choose method for computing gravity component for a cell of a grid}
\usage{
select_gMethod_Edges(xloc, yloc, zloc, gloc, gdiscr, edge, layer, layermax,
  r_inner, r_outer, gamma, rho, w)
}
\arguments{
\item{xloc}{t}

\item{yloc}{t}

\item{zloc}{t}

\item{gloc}{t}

\item{gdiscr}{t}

\item{edge}{t}

\item{layer}{t}

\item{layermax}{t}

\item{r_inner}{t}

\item{r_outer}{t}

\item{gamma}{t}

\item{rho}{t}

\item{w}{t
...}
}
\description{
test
}
\details{
test
}
\examples{
missing 
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
