% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_to_grid3d.R
\name{surface_to_grid3d}
\alias{surface_to_grid3d}
\title{Interpolate grid(df) to 3D grid}
\usage{
surface_to_grid3d(surface_grid, grid_discr, depth_split, Bd_x = Building_x,
  Bd_y = Building_y)
}
\arguments{
\item{grid_discr}{vector of discretization of new grid in (x,y,z).}

\item{Bd_x, y}{Vector, containing min and max coordinates of buildings x and y coordinates.
They only need to be supplied, if no surface grid is given as input.}

\item{grid_cords}{coordinates of the original hydrus mesh.}

\item{data_input}{data to be interpolated. in fomat of the hydrus mesh.}
}
\description{
interpolate DEM to necessary grid of corresponding layer
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2017), mreich@posteo.de
}
